creates component to wrap react-router Link. changes subnav order
