+++ src/components/ListStory.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Link } from 'react-router-dom';
+import Link from 'components/Link';
 import Avatar from 'components/Avatar';
 
 const StoryItemContainer = styled.div`margin-bottom: 2rem;`;
