+++ src/routes/NoMatch.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import Container from 'components/Container';
-import { Link } from 'react-router-dom';
+import Link from 'components/Link';
 
 const Title = styled.h1`margin-top: 25vh;`;
 
@@ -9,7 +9,9 @@ const NoMatch = ({ location }) => {
   return (
     <Container>
       <Title>{`Whoops, ${location.pathname} doesn't exist :(`}</Title>
-      <Link to="/">Head home</Link>
+      <Link to="/" exact>
+        Head home
+      </Link>
     </Container>
   );
 };
